Refactor book handling and improve pagination and filtering logic
- Updated book_params method in BooksController for consistency.
- Fixed export statement in DashboardLayout to ensure proper file structure.
- Enhanced AuthForm to utilize setError for better error handling.
- Modified Login component to pass setError prop for error management.
- Added editingBook prop to BookFormFields for improved form handling.
- Corrected available and total copies display in BookCard component.
- Implemented EmptyState rendering in BookGrid for better user experience.
- Refactored import paths for useBookFilter and usePagination hooks.
- Introduced useBookFilter and usePagination hooks for better code organization and functionality.
