+++ client/src/features/books/components/book-grid/index.js
@@ -3,7 +3,7 @@ export { default as BookFilter } from "./BookFilter";
 export { default as BookCard } from "./BookCard";
 export { default as Pagination } from "./Pagination";
 export { default as EmptyState } from "./EmptyState";
-export { default as useBookFilter } from "./useBookFilter";
-export { default as usePagination } from "./usePagination";
+export { default as useBookFilter } from "../../hooks/useBookFilter";
+export { default as usePagination } from "../../hooks/usePagination";
 export { default as BorrowButton } from "./BorrowButton";
 export { default as ReturnButton } from "./ReturnButton";
